BHASHANET_PROD/requirements.txt
asgiref==3.6.0
autocommand==2.2.2
beautifulsoup4==4.12.2
bs4==0.0.1
captcha==0.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==39.0.2
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.7
django-admin-interface==0.26.0
django-ckeditor==6.7.0
django-colorfield==0.9.0
django-environ==0.10.0
django-import-export==3.1.0
django-js-asset==2.1.0
django-parler==2.3
django-richtextfield==1.6.1
django-rosetta==0.9.8
django-throttle==0.3.1
django-utils-six==2.0
django-validators==1.0.1
dnspython==2.3.0
email-validator==1.3.1
et-xmlfile==1.1.0
gTTS==2.3.1
idna==3.4
json2html==1.3.0
makepass==0.11.0
MarkupPy==1.14
mysql-connector-python==8.1.0
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.1.2
Pillow==9.4.0
polib==1.2.0
protobuf==4.21.12
pycparser==2.21
python-magic==0.4.27
python-slugify==8.0.1
PyYAML==6.0
requests==2.28.2
soupsieve==2.4.1
sqlparse==0.4.3
tablib==3.3.0
text-unidecode==1.3
throttle==0.2.2
tld==0.13
tzdata==2022.7
urllib3==1.26.14
validators==0.20.0
xlrd==2.0.1
xlwt==1.3.0
pymongo==4.6.3
